+++ CHANGELOG.md
@@ -37,6 +37,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [0.19.6]
 
 ### Changed
+- Access log filters config handled
 - Global snapshot auditor added
 - updated project java version 11